fixed the routes persmisions

+++ app/routes.py
@@ -70,12 +70,12 @@ def gconnect():
     
    
 
-    print "hello"
+    
     token =  request.data
     # (Receive token by HTTPS POST)
     # ...
 
-    print token
+   
     try:
         # Specify the CLIENT_ID of the app that accesses the backend:
         idinfo = id_token.verify_oauth2_token(token, requestGoogleAuth, "682221223878-pl3rgk5qvvgme87832b2jeegjejs62og.apps.googleusercontent.com")
@@ -113,87 +113,8 @@ def gconnect():
         # Invalid token
         pass
     
- 
-
-   
-    '''
-    if request.args.get('state') != login_session['state']:
-        response = make_response(json.dumps('Invalid State parameter id'), 401)
-        response.headers['Content-Type'] = 'aplication/json'
-        return response
-    code = request.data
-
-    try:
-        # Updgrade the  auth code to a creddential object
-        oauth_flow = flow_from_clientsecrets('client_secrets.json', scope='')
-        oauth_flow.redirect_uri = 'postmessage'
-        credentials = oauth_flow.step2_exchange(code)
-    except FlowExchangeError:
-        response = make_response(
-            json.dumps('Failed to Upgrade the autho code'), 401)
-        response.headers['Content-Type'] = 'aplication/json'
-        return response
-
-    # check the access token is valid
-    access_token = credentials.access_token
-    url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
-           % access_token)
-    h = httplib2.Http()
-    result = json.loads(h.request(url, 'GET')[1])
-
-    # verification for access token
-    if result.get('error') is not None:
-        response = make_response(json.dumps(result.get('error')), 500)
-        response.headers['Content-Type'] = 'aplication/json'
-        return response
-
-    # Verification for the access toke is used for the  intended user
-    gplus_id = credentials.id_token['sub']
-    if result['user_id'] != gplus_id:
-        response = make_response(
-            json.dumps("Token's user ID does not match giver user ID"), 401
-        )
-        response.headers['Content-Type'] = 'aplication/json'
-        return response
-
-    # Verification for the access token is valid for this app
-    if result['issued_to'] != CLIENT_ID:
-        response = make_response(
-            json.dumps("Token's client ID does not match app"), 401)
-        print"Token's client id does not match app's"
-        response.headers['Content-Type'] = 'application/json'
-        return response
-
-    # check if user is already logged in
-    stored_access_token = login_session.get('access_token')
-    stored_gplus_id = login_session.get('gplus_id')
-    if stored_access_token is not None and gplus_id == stored_gplus_id:
-        response = make_response(
-            json.dumps('Current user is already connected.'), 200)
-        response.headers['Content-Type'] = 'application/json'
-        return response
-
-    # Store the acccess token in the session for later use
-    login_session['access_token'] = credentials.access_token
-    login_session['gplus_id'] = gplus_id
-
-    # Get user info
-    userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
-    params = {'access_token': credentials.access_token, 'alt': 'json'}
-    answer = requests.get(userinfo_url, params=params)
-
-    data = answer.json()
-
-    login_session['username'] = data['name']
-    login_session['email'] = data['email']
-
-    output = login_session['username']
-    createUser()
     
 
-    print "done!"
-    return output
-    '''
 
 @app.route('/gdisconnect', methods=['POST'])
 def gdisconnect():
@@ -202,50 +123,6 @@ def gdisconnect():
     del login_session['email']
     
     return "Success"
-
-    '''
-    access_token = login_session.get('access_token')
-    if access_token is None:
-        print 'Access Token is None'
-        response = make_response(
-            json.dumps('Current user not connected.'), 401)
-        response.headers['Content-Type'] = 'application/json'
-        return response
-    print 'In gdisconnect access token is %s', access_token
-    print 'User name is: '
-    print login_session['username']
-    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' \
-    % login_session['access_token']
-    h = httplib2.Http()
-    result = h.request(url, 'GET')[0]
-    print 'result is '
-    print result
-
-    # If th status is 200 logout the user no problem
-    if result['status'] == '200':
-        print('yeah')
-        del login_session['access_token']
-        del login_session['gplus_id']
-        del login_session['username']
-        del login_session['email']
-        response = make_response(json.dumps('Successfully disconnected.'), 200)
-        response.headers['Content-Type'] = 'application/json'
-        return response
-
-    # If the status is 400 logout the user regardless but printout the problem
-    # to be solve later withou ruining the users experience
-    elif result['status'] == '401':
-        del login_session['access_token']
-        del login_session['gplus_id']
-        del login_session['username']
-        del login_session['email']
-        response = make_response(
-            json.dumps(
-                'Failed to revoke token for given user.',
-                400))
-        response.headers['Content-Type'] = 'application/json'
-        return response
-    '''
 
 # Check if user is currenty signed in
 def usernameState(state):
@@ -257,15 +134,6 @@ def usernameState(state):
         username = False
 
     return username
-
-# creates anti forgery token state
-
-
-
-
-
-    
-
 
 
 # App Errro Handler
@@ -310,12 +178,7 @@ def intruder():
 
 @app.route('/login', methods=['GET','POST'])
 def login():
-    
-    
-  
-  
 
-   
     return render_template('g-login.html', STATE=state)
 
 
@@ -393,22 +256,14 @@ def item_description(category, item):
 
   
     username = usernameState(state)
-
+   
     
 
     category = session.query(Category).filter_by(name=category).one_or_none()
     item = session.query(Item).filter_by(name=item).one_or_none()
-    if username:
-        print "yes"
-        user = session.query(User).filter_by(email=login_session['email']).one_or_none()
-        ids =  user.id == item.user_id
-        return(render_template('item-description.html',
-           item=item, STATE=state, username=username, category=category, id=ids))
+    ids = currentUser(item)
 
-        
 
-  
-    ids = "hello"
     
       
 
@@ -429,6 +284,12 @@ def login_required(f):
             return redirect('/intruder')
     return decorated_function
 
+def currentUser(item):
+    username = usernameState(state)
+    if username:
+        user = session.query(User).filter_by(email=login_session['email']).one_or_none()
+        ids =  user.id == item.user_id
+        return ids
 
 def createUser(name, email):
 
@@ -437,7 +298,8 @@ def createUser(name, email):
     
     u = session.query(User).filter_by(email = email).one_or_none()
     if u == None:
-        # Create ID
+
+        # Create and ID for the user
         id = random.randint(0,9)
         newUser = User(id = id, name=name, email=email)
         session.add(newUser)
@@ -457,28 +319,7 @@ def createUser(name, email):
     else:
         print u.email + str(u.id)
 
-    '''   
-    exists = session.query(User).filter_by(
-        email=login_session['email']).scalar() is not None
-    if exists:
-        return session.query(User).filter_by(email = login_session['email'])
-    print session.query(User).all()
-    print login_session
-    
-    sessionQuery =  session.query(User).delete()
-   
-    print user.email    
-    if user.email == loggin_session["email"]:
-        print "yes"
-
-
-    print login_session
-    newUser = User(name=login_session['username'], email=login_session[
-                   'email'])
-    session.add(newUser)
-    session.commit()
-    return user.id
-    '''
+ 
 
 def getUserInfo(user_id):
     user = session.query(User).filter_by(id=user_id).one_or_none()
@@ -528,33 +369,34 @@ def edit(category, item):
 
    
     username = usernameState(state)
-    
-    
 
     category = session.query(Category).all()
     item = session.query(Item).filter_by(name=item).one_or_none()
-   
 
-   
+    if currentUser(item):
+        print "yes"
 
-    if request.method == 'POST':
-        if request.form['name']:
-            item.name = request.form['name']
-        if request.form['description']:
-            item.description = request.form['description']
-        if request.form['categories']:
-            item.category_id = request.form['categories']
 
-        session.add(item)
-        session.commit()
+        if request.method == 'POST':
+            if request.form['name']:
+                item.name = request.form['name']
+            if request.form['description']:
+                item.description = request.form['description']
+            if request.form['categories']:
+                item.category_id = request.form['categories']
 
-        return redirect(url_for('index'))
-    return render_template(
-        'edit.html',
-        item=item,
-        category=category,
-        username=username,
-        STATE=state)
+            session.add(item)
+            session.commit()
+
+            return redirect(url_for('index'))
+        return render_template(
+            'edit.html',
+            item=item,
+            category=category,
+            username=username,
+            STATE=state)
+    else:
+        return "Sorry you are not allow to edit post you have not made yourself return <a href='/'>home</a>"   
 
 @app.route('/catalog/<category>/<item>/delete', methods=['GET', 'POST'])
 @login_required
@@ -565,15 +407,19 @@ def delete(category, item):
 
     item = session.query(Item).filter_by(name=item).one_or_none()
     category = session.query(Category).filter_by(name=category).one_or_none()
+    
+    if currentUser(item):
 
-    if request.method == 'POST':
-        session.delete(item)
-        session.commit()
-        return redirect('/')
+        if request.method == 'POST':
+            session.delete(item)
+            session.commit()
+            return redirect('/')
 
-    return render_template(
-        'delete.html',
-        item=item,
-        category=category,
-        STATE=state,
-        username=username)
+        return render_template(
+            'delete.html',
+            item=item,
+            category=category,
+            STATE=state,
+            username=username)
+    else:
+        return "sorry you are not allow to delete items you have not created, return  <a href='/'> home</a>"
